Move FragmentToBatch to separate class
