+++ Detectors/PHOS/workflow/src/StandaloneAODProducerSpec.cxx
@@ -19,7 +19,6 @@
 #include "Framework/InputRecordWalker.h"
 #include "Framework/Logger.h"
 #include "Framework/TableBuilder.h"
-#include "Framework/TableTreeHelpers.h"
 #include "MathUtils/Utils.h"
 
 using namespace o2::framework;
@@ -106,7 +105,7 @@ void StandaloneAODProducerSpec::run(ProcessingContext& pc)
                       o2::math_utils::detail::truncateFloatFraction(c.getTime(), mCaloTime),
                       c.getType(), // HG/LG
                       0);          // hard coded for phos (-1 would be undefined, 0 phos)
-    }                              // end of cell loop
+    } // end of cell loop
 
     auto bcID = tr.getBCData().toLong();
     bcCursor(0,
