+++ Framework/AnalysisSupport/src/DataInputDirector.cxx
@@ -16,8 +16,8 @@
 #include "Framework/AnalysisDataModelHelpers.h"
 #include "Framework/Output.h"
 #include "Framework/Signpost.h"
+#include "Framework/FragmentToBatch.h"
 #include "Headers/DataHeader.h"
-#include "Framework/TableTreeHelpers.h"
 #include "Monitoring/Tags.h"
 #include "Monitoring/Metric.h"
 #include "Monitoring/Monitoring.h"
