+++ Framework/Core/src/DataAllocator.cxx
@@ -11,7 +11,6 @@
 #include "Framework/CompilerBuiltins.h"
 #include "Framework/Lifetime.h"
 #include "Framework/TableBuilder.h"
-#include "Framework/TableTreeHelpers.h"
 #include "Framework/DataAllocator.h"
 #include "Framework/MessageContext.h"
 #include "Framework/ArrowContext.h"
@@ -19,6 +18,7 @@
 #include "Framework/DataProcessingHeader.h"
 #include "Framework/FairMQResizableBuffer.h"
 #include "Framework/DataProcessingContext.h"
+#include "Framework/FragmentToBatch.h"
 #include "Framework/DeviceSpec.h"
 #include "Framework/StreamContext.h"
 #include "Framework/Signpost.h"
